장병문 : 인스턴싱 준비중 업로드, 파이프라인 -> 마테리얼로 변경작업 업로드

+++ Project/GameEngineCore/GameEngineFBXAnimationRenderer.cpp
@@ -335,10 +335,10 @@ void GameEngineFBXAnimationRenderer::ChangeAnimation(const std::string& _Animati
 }
 
 
-void GameEngineFBXAnimationRenderer::ChangeAnimationFrame( int _StartIndex)
+void GameEngineFBXAnimationRenderer::ChangeAnimationFrame(size_t _StartIndex)
 {
-	int ChangeFrame = _StartIndex;
-	int MaxSize = GetAnimationFrameCount();
+	size_t ChangeFrame = _StartIndex;
+	size_t MaxSize = GetAnimationFrameCount();
 	if (MaxSize <= _StartIndex)
 	{
 		ChangeFrame = MaxSize;
